backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
numpy==1.26.4
python-multipart==0.0.9
pillow==10.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
supabase==2.3.0
asyncpg==0.29.0

# AI Engine
google-generativeai==0.3.2
openai==1.12.0

# Background Tasks
arq==0.25.0
redis==5.0.1

# WebSockets
websockets==12.0

# Configuration & Utils
python-dotenv==1.0.0
structlog==24.1.0
httpx==0.26.0
tenacity==8.2.3

# Monitoring
prometheus-client==0.19.0
sentry-sdk==1.40.0
